Allow overriding all expr by moving other_expr to the top
