+++ core/src/main/kotlin/com/alecstrong/sql/psi/core/sql.bnf
@@ -201,7 +201,8 @@ drop_view_stmt ::= DROP VIEW [ IF EXISTS ] [ database_name DOT ] view_name {
   stubClass = "com.alecstrong.sql.psi.core.psi.SchemaContributorStub"
   pin = 2
 }
-expr ::= ( raise_expr
+expr ::= ( other_expr
+         | raise_expr
          | case_expr
          | exists_expr
          | multi_column_expr
@@ -225,7 +226,6 @@ expr ::= ( raise_expr
          | binary_pipe_expr
          | unary_expr
          | literal_expr
-         | other_expr
          | column_expr
          | bind_expr )
