--- sample-headless/src/main/kotlin/com/alecstrong/sqlite/psi/sample/headless/SampleHeadlessParser.kt
@@ -1,30 +1,0 @@
-package com.alecstrong.sql.psi.sample.headless
-
-import com.alecstrong.sql.psi.core.SqlCoreEnvironment
-import com.alecstrong.sql.psi.sample.core.SampleFileType
-import com.alecstrong.sql.psi.sample.core.SampleParserDefinition
-import java.io.File
-
-class SampleHeadlessParser {
-  fun parseSqlite() {
-    val parserDefinition = SampleParserDefinition()
-    val environment = object : SqlCoreEnvironment(
-      sourceFolders = listOf(File("sample-headless")),
-      dependencies = emptyList(),
-    ) {
-      init {
-        initializeApplication {
-          registerFileType(SampleFileType, SampleFileType.defaultExtension)
-          registerParserDefinition(parserDefinition)
-        }
-      }
-    }
-    environment.annotate { _, message ->
-      System.err.println(message)
-    }
-  }
-}
-
-fun main() {
-  SampleHeadlessParser().parseSqlite()
-}
