--- sample-headless/src/main/kotlin/com/alecstrong/sqlite/psi/sample/headless/test.samplesql
@@ -1,7 +1,0 @@
-CREATE TABLE test (
-  sample_column TEXT AS "java.util.List" NOT NULL
-);
-
-SELECT *
-FROM test
-WHERE sample_column = fake_column;
